[Refactor] Group hub:sync events
Also move function bodies to hub-utils
